+++ packages/flutter_login_firebase/lib/src/firebase_login.dart
@@ -1,60 +1,157 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 // SPDX-FileCopyrightText: 2022 Iconica
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter_login_interface/flutter_login_interface.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_login_service/flutter_login_service.dart';
 
-class FirebaseLogin extends LoginInterface<User> {
-  final FirebaseApp app;
+/// A service to handle authentication with Firebase.
+/// Takes an instance of [FirebaseAuth] and an
+/// optional [LoginServiceTranslations]
+class FirebaseLoginService implements LoginService {
+  FirebaseLoginService({
+    required this.auth,
+    this.translationsService = const LoginServiceTranslations.empty(),
+  });
 
-  FirebaseLogin(this.app);
+  /// The [FirebaseAuth] instance to use for authentication.
+  final FirebaseAuth auth;
 
+  /// The translations service to use for error messages.
+  final LoginServiceTranslations translationsService;
+
+  /// The error response from the last login attempt.
+  LoginResponse? loginError;
+
+  /// The error response from the last password reset request.
+  RequestPasswordResponse? requestPasswordError;
+
+  /// Get the currently logged in user.
   @override
-  Future<LoginResponse<User>> loginWithEmailAndPassword(
-      EmailPasswordLogin login,
-      {Function(dynamic resolver)? onMFA}) async {
-    var auth = FirebaseAuth.instanceFor(app: app);
+  Future<User?> getLoggedInUser() async => auth.authStateChanges().first;
+
+  /// Log in with email and password.
+  /// Returns a [LoginResponse] with the user object if the login was
+  /// successful. If the login was not successful, the response will
+  /// contain an [Error] object with the error title and message.
+  @override
+  Future<LoginResponse> loginWithEmailAndPassword(
+    String email,
+    String password,
+    BuildContext context, {
+    // ignore: avoid_annotating_with_dynamic
+    Function(dynamic resolver)? onMFA,
+  }) async {
     try {
       var credential = await auth.signInWithEmailAndPassword(
-        email: login.email,
-        password: login.password,
+        email: email,
+        password: password,
       );
       return LoginResponse(
         loginSuccessful: credential.user != null,
         userObject: credential.user,
       );
-    } on FirebaseAuthException catch (_) {
-      return LoginResponse(loginSuccessful: false, userObject: null);
+    } on FirebaseAuthException catch (e) {
+      loginError = switch (e.code) {
+        'user-disabled' => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.userDisabledMessage,
+              message: translationsService.userDisabledMessage,
+            ),
+          ),
+        'wrong-password' => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.wrongPasswordTitle,
+              message: translationsService.wrongPasswordMessage,
+            ),
+          ),
+        'too-many-requests' => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.tooManyRequestsTitle,
+              message: translationsService.tooManyRequestsMessage,
+            ),
+          ),
+        'invalid-credential' => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.invalidCredentialsTitle,
+              message: translationsService.invalidCredentialsMessage,
+            ),
+          ),
+        _ => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.errorMessage,
+              message: translationsService.errorMessage,
+            ),
+          ),
+      };
     }
+    return loginError!;
   }
 
+  /// Log out the currently logged in user.
+  /// Returns `true` if the logout was successful, `false` if it was not.
   @override
-  Future<bool> requestPasswordReset(String email) async {
-    var auth = FirebaseAuth.instanceFor(app: app);
-    try {
-      await auth.sendPasswordResetEmail(email: email);
-      return true;
-    } catch (_) {
-      return false;
-    }
-  }
-
-  @override
-  Future<User?> getLoggedInUser() {
-    var auth = FirebaseAuth.instanceFor(app: app);
-    return auth.authStateChanges().first;
-  }
-
-  @override
-  Future<bool> logout() async {
-    var auth = FirebaseAuth.instanceFor(app: app);
+  Future<bool> logout(
+    BuildContext context,
+  ) async {
     try {
       await auth.signOut();
       return true;
-    } catch (_) {
+    } on FirebaseAuthException catch (_) {
       return false;
+    }
+  }
+
+  /// Request a password reset for the given email.
+  /// Returns a [RequestPasswordResponse] with a boolean indicating
+  /// whether the request was successful. If the request was not
+  /// successful, the response will contain an [Error] object with
+  /// the error title and message.
+  @override
+  Future<RequestPasswordResponse> requestChangePassword(
+    String email,
+    BuildContext context,
+  ) async {
+    try {
+      await auth.sendPasswordResetEmail(email: email);
+      return const RequestPasswordResponse(requestSuccesfull: true);
+    } on FirebaseAuthException catch (e) {
+      requestPasswordError = switch (e.code) {
+        'invalid-email' => RequestPasswordResponse(
+            requestSuccesfull: false,
+            requestPasswordError: Error(
+              title: translationsService.changePasswordInvalidEmailTitle,
+              message: translationsService.changePasswordInvalidEmailMessage,
+            ),
+          ),
+        'user-not-found' => RequestPasswordResponse(
+            requestSuccesfull: false,
+            requestPasswordError: Error(
+              title: translationsService.changePasswordUserNotFoundTitle,
+              message: translationsService.changePasswordUserNotFoundMessage,
+            ),
+          ),
+        _ => RequestPasswordResponse(
+            requestSuccesfull: false,
+            requestPasswordError: Error(
+              title: translationsService.changePasswordErrorTitle,
+              message: translationsService.changePasswordErrorMessage,
+            ),
+          ),
+      };
+      return requestPasswordError!;
     }
   }
 }