+++ packages/flutter_login_shared_preferences/lib/flutter_login_shared_preferences.dart
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2022 Iconica
-//
-// SPDX-License-Identifier: BSD-3-Clause
-
-library flutter_login_shared_preferences;
-
-export 'src/shared_preferences_data_source.dart';