+++ packages/flutter_login_firebase/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* Removed the `LoginServiceInterface` class.
+* Added Responses to the `logInWithEmailAndPassword` and `requestChangePassword`.
+
 ## 1.0.0
 
 * Updated all dependencies to latest versions.