+++ packages/flutter_login_service/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 
+
+* Removed the `LoginServiceInterface` class.
+* Added `RequestPasswordResponse`.
+* Added an `Error` object to the `LoginResponse` object.
+
 ## 1.0.0
 
 * Updated all dependencies to latest versions.